move JavaEvaluator to ml.evaluation as well
